Added 2 plotly charts
