+++ templates/plotly_titanic.html
