+++ templates/scores.html
@@ -25,7 +25,7 @@
       <tbody>
         {% for score in scores.underage10 %}
         <tr>
-          <th scope="row">{% if score.rank == 1 %}<img class="crown" alt="golden crown" src="{{ url_for('static', filename='/img/crown_gold.png') }}">{% elif score.rank == 2 %}<img class="crown" alt="silver crown" src="{{ url_for('static', filename='/img/crown_silver.png') }}">{% elif score.rank == 3 %}<img class="crown" alt="bronze crown" src="{{ url_for('static', filename='/img/crown_bronze.png') }}">{% else %}{{ score.rank }}{% endif %}</th>
+          <th scope="row">{% if score.rank == 1 %}<img class="crown" alt="golden crown" src="{{ url_for('static', filename='img/crown_gold.png') }}">{% elif score.rank == 2 %}<img class="crown" alt="silver crown" src="{{ url_for('static', filename='img/crown_silver.png') }}">{% elif score.rank == 3 %}<img class="crown" alt="bronze crown" src="{{ url_for('static', filename='img/crown_bronze.png') }}">{% else %}{{ score.rank }}{% endif %}</th>
           <td>
             {% if score.shared %}
 			<a href="javascript:void(0);" class="dotted" tabindex="0" data-bs-toggle="popover" data-bs-trigger="focus" data-bs-html="true" data-bs-content="<b>Gender:</b> {{ score.gender }}<br><b>Birthdate:</b> {{ score.birthdate }}<br><b>Education:</b> {{ score.education }}<br><i>{{ score.bio }}</i>">
@@ -57,7 +57,7 @@
       <tbody>
         {% for score in scores.overall %}
         <tr>
-          <th scope="row">{% if score.rank == 1 %}<img class="crown" alt="golden crown" src="{{ url_for('static', filename='/img/crown_gold.png') }}">{% elif score.rank == 2 %}<img class="crown" alt="silver crown" src="{{ url_for('static', filename='/img/crown_silver.png') }}">{% elif score.rank == 3 %}<img class="crown" alt="bronze crown" src="{{ url_for('static', filename='/img/crown_bronze.png') }}">{% else %}{{ score.rank }}{% endif %}</th>
+          <th scope="row">{% if score.rank == 1 %}<img class="crown" alt="golden crown" src="{{ url_for('static', filename='img/crown_gold.png') }}">{% elif score.rank == 2 %}<img class="crown" alt="silver crown" src="{{ url_for('static', filename='img/crown_silver.png') }}">{% elif score.rank == 3 %}<img class="crown" alt="bronze crown" src="{{ url_for('static', filename='img/crown_bronze.png') }}">{% else %}{{ score.rank }}{% endif %}</th>
           <td>
             {% if score.shared %}
 			<a href="javascript:void(0);" class="dotted" tabindex="0" data-bs-toggle="popover" data-bs-trigger="focus" data-bs-html="true" data-bs-content="<b>Gender:</b> {{ score.gender }}<br><b>Birthdate:</b> {{ score.birthdate }}<br><b>Education:</b> {{ score.education }}<br><i>{{ score.bio }}</i>">
@@ -89,7 +89,7 @@
       <tbody>
         {% for score in scores.avgtime %}
         <tr>
-          <th scope="row">{% if score.rank == 1 %}<img class="crown" class="crown" alt="golden crown" src="{{ url_for('static', filename='/img/crown_gold.png') }}">{% elif score.rank == 2 %}<img class="crown" alt="silver crown" src="{{ url_for('static', filename='/img/crown_silver.png') }}">{% elif score.rank == 3 %}<img class="crown" alt="bronze crown" src="{{ url_for('static', filename='/img/crown_bronze.png') }}">{% else %}{{ score.rank }}{% endif %}</th>
+          <th scope="row">{% if score.rank == 1 %}<img class="crown" class="crown" alt="golden crown" src="{{ url_for('static', filename='img/crown_gold.png') }}">{% elif score.rank == 2 %}<img class="crown" alt="silver crown" src="{{ url_for('static', filename='img/crown_silver.png') }}">{% elif score.rank == 3 %}<img class="crown" alt="bronze crown" src="{{ url_for('static', filename='img/crown_bronze.png') }}">{% else %}{{ score.rank }}{% endif %}</th>
           <td>
             {% if score.shared %} 
 			<a href="javascript:void(0);" class="dotted" tabindex="0" data-bs-toggle="popover" data-bs-trigger="focus" data-bs-html="true" data-bs-content="<b>Gender:</b> {{ score.gender }}<br><b>Birthdate:</b> {{ score.birthdate }}<br><b>Education:</b> {{ score.education }}<br><i>{{ score.bio }}</i>">
