+++ templates/semenov.html
@@ -58,7 +58,7 @@
 				<div class="row row-cols-1 row-cols-md-3 mt-5 mb-5">
 				  <div class="col mt-1 mb-1">
 					<div class="card h-100">
-					  <img src="{{ url_for('static', filename='/img/me_Dali_sm.jpg') }}" class="card-img-top" alt="Gala Dali Castle House-Museum">
+					  <img src="{{ url_for('static', filename='img/me_Dali_sm.jpg') }}" class="card-img-top" alt="Gala Dali Castle House-Museum">
 					  <div class="card-body">
 						<p class="card-text">I am in the museum of Salvador Dali.<br>Gala Dali Castle House-Museum</p>
 					  </div>
@@ -66,7 +66,7 @@
 				  </div>
 				  <div class="col mt-1 mb-1">
 					<div class="card h-100">
-					  <img src="{{ url_for('static', filename='/img/family_sm.jpg') }}" class="card-img-top" alt="The Lighthouse On The Punta Del Albir">
+					  <img src="{{ url_for('static', filename='img/family_sm.jpg') }}" class="card-img-top" alt="The Lighthouse On The Punta Del Albir">
 					  <div class="card-body">
 						<p class="card-text">My family on a NY trip to Spain.<br>The Lighthouse On The Punta Del Albir.</p>
 					  </div>
@@ -74,7 +74,7 @@
 				  </div>
 				  <div class="col mt-1 mb-1">
 					<div class="card h-100">
-					  <img src="{{ url_for('static', filename='/img/me_Levante_sm.jpg') }}" class="card-img-top" alt="Levante view, Benidorm">
+					  <img src="{{ url_for('static', filename='img/me_Levante_sm.jpg') }}" class="card-img-top" alt="Levante view, Benidorm">
 					  <div class="card-body">
 						<p class="card-text">I love hiking.<br>Benidorm's Levante view from high point.</p>
 					  </div>
