--- webApp/README.md
@@ -1,1 +1,0 @@
-# my_site
