--- webApp/requirements.txt
@@ -1,7 +1,0 @@
-flask
-flask-session
-flask-mail
-tempfile
-functools
-urllib
-sqlite3
