--- webApp/static/favicon/favicon_about.txt
@@ -1,6 +1,0 @@
-This favicon was generated using the following font:
-
-- Font Title: Amaranth
-- Font Author: Copyright 2016 The Amaranth Project Authors (https://github.com/googlefonts/amaranth)
-- Font Source: http://fonts.gstatic.com/s/amaranth/v11/KtkuALODe433f0j1zPnCF9GqwnzW.ttf
-- Font License: SIL Open Font License, 1.1 (http://scripts.sil.org/OFL)
