--- webApp/static/favicon/site.webmanifest
@@ -1,1 +1,0 @@
-{"name":"Project","short_name":"Proj","icons":[{"src":"android-chrome-192x192.png","sizes":"192x192","type":"image/png"},{"src":"android-chrome-512x512.png","sizes":"512x512","type":"image/png"}],"theme_color":"#ffffff","background_color":"#ffffff","display":"standalone"}
