--- webApp/static/index.css
@@ -1,9 +1,0 @@
-div #qname
-{
-    display: none;
-}
-div.inline *
-{
-    display: inline;
-}
-
