--- webApp/static/results.css
@@ -1,7 +1,0 @@
-td.right {
-    color:#1b1;
-}
-
-td.wrong {
-    color:#f00;
-}
