--- webApp/static/scores.css
@@ -1,3 +1,0 @@
-.crown {
-    height: 24px;
-}
