--- webApp/static/scripts.js
@@ -1,10 +1,0 @@
-function sleep(ms) {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-function listener() {
-    // Dynamic notifications system
-    
-}
-
-document.addEventListener('DOMContentLoaded', listener);
