--- webApp/static/styles.css
@@ -1,63 +1,0 @@
-nav {
-    padding-left: 1rem;
-    padding-right: 1rem;
-}
-nav .navbar-brand
-{
-    /* size for brand */
-    font-size: xx-large;
-}
-
-/* Colors for brand */
-nav .navbar-brand .blue
-{
-    color: #537fbe;
-}
-nav .navbar-brand .red
-{
-    color: #ea433b;
-}
-nav .navbar-brand .yellow
-{
-    color: #f5b82e;
-}
-nav .navbar-brand .green
-{
-    color: #2e944b;
-}
-
-nav .ml-auto 
-{
-    margin-left: auto;
-}
-
-main
-{
-    /* Scroll horizontally as needed */
-    overflow-x: auto;
-
-    /* Center contents */
-    text-align: center;
-}
-
-main img
-{
-    /* Constrain images on small screens */
-    max-width: 100%;
-}
-
-main table {
-    /*text-align: left;*/
-}
-
-
-#notifications {
-  position: fixed;
-  width: 100%;
-  z-index: 7000;
-  top: 0;
-}
-
-.dotted {
-  text-decoration: 1px underline dotted;
-}
