--- webApp/static/test_generate.js
@@ -1,42 +1,0 @@
-function listener() {
-
-    const notifications = document.getElementById('notifications');
-    rebindSubmit(document.getElementById('formGenerate'));
-    
-    function update(myForm, lastInput) {
-        const xhr = new XMLHttpRequest();
-
-        xhr.onload = function() {
-          if (xhr.readyState === xhr.DONE && xhr.status === 200) {
-            notifications.insertAdjacentHTML('afterbegin', xhr.response);
-            alertTimer();
-            // Bootstrap's button focus fix
-            lastInput.focus(); lastInput.blur();
-              }
-          }
-
-        xhr.onerror = function() {
-          dump("Error while getting xhr.");
-        }
-    
-        xhr.open(myForm.method, myForm.action, true);
-        xhr.responseType = "json";
-        xhr.send(new FormData(myForm));
-    }
-
-    function rebindSubmit(submitForm){
-        submitForm.addEventListener('submit', function(evt) {
-            evt.preventDefault();
-            update(submitForm, submitForm.elements[submitForm.length-2]);
-        });
-    }
-        
-    function alertTimer() {
-        $(".alert").first().hide().slideDown(500).delay(4000).slideUp(200, function(){
-            $(this).remove(); 
-        });
-    }
-}
-
-
-document.addEventListener('DOMContentLoaded', listener);
