--- webApp/templates/apology.html
@@ -1,7 +1,0 @@
-{% extends "layout.html" %}
-
-{% block title %}Apology{% endblock %}
-
-{% block main %}
-    <img alt="{{ top }}" class="border" src="http://memegen.link/custom/{{ top | urlencode }}/{{ bottom | urlencode }}.jpg?alt=https://i.imgur.com/CsCgN7Ll.png" title="{{ top }}">
-{% endblock %}
