--- webApp/templates/index.html
@@ -1,27 +1,0 @@
-{% extends "layout.html" %}
-{% block head %}
-    <script type="text/javascript" src="{{ url_for('static', filename='index.js') }}"></script>
-    <link href="{{ url_for('static', filename='index.css') }}" rel="stylesheet">
-{% endblock %}
-{% block title %}Main page{% endblock %}
-
-{% block main %}
-    <div class="container-fluid">
-        <div class="row">
-            <h1>This is my CS50 final project</h1>
-        </div>
-        
-        <div class="row">
-        
-            <h2>Hello, <span id="userName">{{ session.user_name }}</span>!</h2>
-            {% if not session.user_name %}
-            <div class="alert alert-secondary alert-dismissible fade show" role="alert"><strong>Oops!</strong> Looks like I've missed your <a href="#" class="alert-link">name</a><div id="qname" class="inline"><form id="formNameUpdate" class="form-group"><input name="name" autocomplete="off" placeholder="Name" type="text" class="input-group-sm"></form></div>!<button type="button" class="btn-close" data-bs-dismiss="alert"></button></div>
-            {% endif %}
-            <div class="container-fluid mt-5 mb-5">
-                <p>This site was made from scratch with knowledge I have learned from Harvard's Computer Science course: <a href="https://cs50.harvard.edu/">CS50</a><br>
-                My special thanks to <a href="https://cs.harvard.edu/malan/">David J. Malan</a>, Gordon McKay Professor of the Practice of Computer Science</p>
-            </div>
-        </div>
-        
-    </div>
-{% endblock %}
