--- webApp/templates/login.html
@@ -1,25 +1,0 @@
-{% extends "layout.html" %}
-{% block title %}Log In{% endblock %}
-{% block main %}
-    <div class="container-fluid mb-5">
-        <form action="/login" method="post">
-            <div class="row row-cols-lg-auto g-1 justify-content-center">
-                <div class="col-12">
-                    <div class="input-group">
-                        <label class="input-group-text">Lgn</label>
-                        <input type="text" class="form-control" autofocus autocomplete="off" name="login" placeholder="Username">
-                    </div>
-                </div>
-                <div class="col-12">
-                    <div class="input-group">
-                        <label class="input-group-text">Pswd</label>
-                        <input type="password" class="form-control" autocomplete="off" name="password" placeholder="***">
-                    </div>
-                </div>
-                <div class="col-12">
-                    <button type="submit" class="btn btn-primary">Log In</button>
-                </div>
-            </div>
-        </form>
-    </div>
-{% endblock %}
