--- webApp/templates/profile.html
@@ -1,113 +1,0 @@
-{% extends "layout.html" %}
-{% block head %}
-<script type="text/javascript" src="{{ url_for('static', filename='profile.js') }}"></script>
-<link href={{ url_for( 'static', filename='profile.css' ) }} rel="stylesheet">
-{% endblock %}
-{% block title %}Profile{% endblock %}
-{% block main %}
-<h2>Profile:</h2>
-<p>The more you tell about yourself, the more accurate and precise overall statistics will be.</p>
-<div class="container-fluid mb-5">
-  <form action="/profile_update" method="post" id="formProfileUpdate">
-    <div class="row row-cols-lg-4 g-1 mb-1 justify-content-center">
-      <div class="col-lg">
-        <div class="input-group">
-          <label class="input-group-text">Gender</label>
-          <select class="form-select" name="gender">
-            {% if not profile.gender %}
-            <option disabled selected value="" />
-			{% endif %}
-			{% for gender in profile.gender_list %}
-            <option value="{{ gender }}" {% if profile.gender==gender %}selected{% endif %}>{{ gender }}</option>
-			{% endfor %}
-          </select>
-        </div>
-      </div>
-      <div class="col-lg">
-        <div class="input-group">
-          <label class="input-group-text">Birthdate</label>
-          <input class="form-control" name="birthdate" value="{% if profile.birthdate %}{{ profile.birthdate }}{% endif %}" type="date" />
-        </div>
-      </div>
-    </div>
-    <div class="row row-cols-lg-2 g-1 mb-1 justify-content-center">
-      <div class="col-lg">
-        <div class="input-group">
-          <label class="input-group-text">Education</label>
-          <input class="form-control" name="education" autocomplete="off" placeholder="Education" value="{% if profile.education %}{{ profile.education }}{% endif %}" type="text">
-        </div>
-      </div>
-    </div>
-    <div class="row row-cols-lg-auto g-1 mb-1 justify-content-center">
-      <div class="col-lg-6">
-        <div class="form-floating">
-          <textarea class="form-control" name="bio" placeholder="">{% if profile.bio %}{{ profile.bio }}{% endif %}</textarea>
-          <label>What do you like?</label>
-        </div>
-      </div>
-    </div>
-    <div class="row row-cols-lg-auto g-1 mb-1 justify-content-center">
-      <div class="col-lg">
-        <div class="form-check">
-          <label class="form-check-label">
-            <input class="form-check-input" type="checkbox" id="checkShared" value="{{ profile.shared }}" name="shared" {% if profile.shared %}checked{% endif %}>Open profile for public
-          </label>
-        </div>
-      </div>
-    </div>
-    <div class="row row-cols-lg-auto g-1 justify-content-center">
-      <div class="col-lg">
-        <button class="btn btn-primary" type="submit">Update</button>
-      </div>
-    </div>
-  </form>
-</div>
-<div class="container-fluid mb-5">
-  <p>To change email:</p>
-  <form class="row row-cols-lg-auto g-1 justify-content-center" action="/email_update" method="post" id="formEmailUpdate">
-    <div class="col-lg-3">
-      <div class="input-group">
-        <label class="input-group-text">Email</label>
-        <input class="form-control" name="email" autocomplete="off" placeholder="What's your email?" value="{% if profile.email %}{{ profile.email }}{% endif %}" type="email">
-      </div>
-    </div>
-    <div class="col-lg">
-      <button class="btn btn-primary" type="submit">Change</button>
-    </div>
-  </form>
-</div>
-<div class="container-fluid mb-5">
-  <p>To change name:</p>
-  <form class="row row-cols-lg-auto g-1 justify-content-center" action="/name_update" method="post" id="formNameUpdate">
-    <div class="col-lg-3">
-      <div class="input-group">
-        <label class="input-group-text">Name</label>
-        <input class="form-control" name="name" autocomplete="off" placeholder="What's your name?" value="{% if profile.name %}{{ profile.name }}{% endif %}" type="text">
-      </div>
-    </div>
-    <div class="col-lg">
-      <button class="btn btn-primary" type="submit">Change</button>
-    </div>
-  </form>
-</div>
-<div class="container-fluid mb-5">
-  <p>To change password:</p>
-  <form class="row row-cols-lg-auto g-1 justify-content-center" action="/password_update" method="post" id="formPasswordUpdate">
-    <div class="col-lg-3">
-      <div class="input-group">
-        <label class="input-group-text">Pswd</label>
-        <input class="form-control" name="password" placeholder="New password" type="password">
-      </div>
-    </div>
-    <div class="col-lg-3">
-      <div class="input-group">
-        <label class="input-group-text">Cnfrm</label>
-        <input class="form-control" name="confirmation" autocomplete="off" placeholder="New password (again)" type="password">
-      </div>
-    </div>
-    <div class="col-lg">
-      <button class="btn btn-primary" type="submit">Change</button>
-    </div>
-  </form>
-</div>
-{% endblock %}
