--- webApp/templates/test_generate.html
@@ -1,24 +1,0 @@
-{% extends "layout.html" %}
-{% block head %}
-    <script type="text/javascript" src="{{ url_for('static', filename='test_generate.js') }}"></script>
-{% endblock %}
-{% block title %}Generate tests{% endblock %}
-
-{% block main %}
-    <h2>Generate tests:</h2>
-    <div class="container">
-        <form action="/test_generate" method="post" id="formGenerate">
-            <div class="row row-cols-lg-auto g-1 mt-5 mb-5 justify-content-center">
-                <div class="col-lg-3">
-                    <div class="input-group">
-                        <div class="input-group-text">Examples</div>
-                        <input class="form-control" name="examples" autocomplete="off" placeholder="Num of examples" type="number" value="20" min="0" max="99">
-                    </div>
-                </div>
-                <div class="col-lg-0">
-                    <button class="btn-primary form-control" type="submit">Generate</button>
-                </div>
-            </div>
-        </form>
-    </div>
-{% endblock %}
