--- webApp/templates/tests.html
@@ -1,79 +1,0 @@
-{% extends "layout.html" %}
-{% block head %}
-    <script type="text/javascript" src="{{ url_for('static', filename='tests.js') }}"></script>
-    <link href={{ url_for('static', filename='tests.css') }} rel="stylesheet">
-{% endblock %}
-{% block title %}Tests{% endblock %}
-
-{% block main %}
-    <h2>Tests:</h2>
-    <h3>Are you ready for a quick and easy math test?</h3>
-    <p>Measuring your reaction, accuracy and precision</p>
-    <div class="container">
-        <div class="container mt-5 mb-5" id="testQuery">
-            {% if test_continue %}
-            <h4>Unfinished tests</h4>
-            <form action="/test_continue" method="get" id="testContinue">
-                <div class="row row-cols-lg-auto g-1 justify-content-center">
-                    <div class="col-lg">
-                        <div class="input-group">
-                            <label class="input-group-text">Test</label>
-                            <select class="form-select" name="test_id" required>
-                                {% for test in test_continue %}
-                                <option value="{{ test.id }}">id: [{{ test.id }}] date: [{{ test.date }}] levels: [{{ test.levels }}] operators: [{{ test.operators }}] examples: [{{ test.examples }}]</option>
-                                {% endfor %}
-                            </select>
-                        </div>
-                    </div>
-                    <div class="col-lg-0">
-                        <button class="btn btn-primary" type="submit">Continue!</button>
-                    </div>
-                </div>
-            </form>
-            {% endif %}
-            <h5 class="mt-5">Start a new one</h5>
-            <form action="/test_start" method="get" id="testStart">
-                <div class="row row-cols-lg-auto g-1 justify-content-center">
-                    <div class="col-lg-2">
-                        <div class="form-floating">
-                            <select class="form-select h-100" multiple name="level" required>
-                                {% for level in levels %}
-                                <option {% if level.level == 1 %}selected {% endif %}value="{{ level.level }}">Level: {{ level.level }}</option>
-                                {% endfor %}
-                            </select>
-                            <label>Pick level(s):</label>
-                        </div>
-                    </div>
-                    <div class="col-lg-2">
-                        <div class="form-floating">
-                            <select class="form-select h-100" multiple name="operator" required>
-                                {% for operator in operators %}
-                                <option selected value="{{ operator.operator }}">Operation: {{ operator.operator }}</option>
-                                {% endfor %}
-                            </select>
-                            <label>Pick operation(s):</label>
-                        </div>
-                    </div>
-                    <div class="col-lg-2">
-                        <div class="input-group">
-                            <label class="input-group-text">Examples</label>
-                            <input class="form-control" name="examples" autocomplete="off" placeholder="Num of examples" type="number" value="10" min="1" max="100">
-                        </div>
-                    </div>
-                    <!--
-                    <div class="col-lg-3">
-                        <div class="input-group">
-                            <div class="input-group-text">Time (ms)</div>
-                            <input class="form-control" name="time" autocomplete="off" placeholder="Time (milliseconds)" type="number" value="10000" min="1000" max="100000">
-                        </div>
-                    </div>
-                    -->
-                    <div class="col-lg-0">
-                        <button class="btn btn-primary" type="submit">I'm ready!</button>
-                    </div>
-                </div>
-            </form>
-        </div>
-        <div class="container" id="test"></div>
-    </div>
-{% endblock %}
